requirements.txt
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
xgboost==2.0.3
hmmlearn==0.3.2
ta==0.11.0
fastapi==0.115.12
uvicorn==0.30.1
plotly==5.22.0
joblib==1.4.2
scipy==1.13.0
pyarrow==16.1.0
requests>=2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
